# Change "value" package altogether to avoid copying everything