--- a/fieldpath/serialize-pe.go
+++ b/fieldpath/serialize-pe.go
@@ -83,14 +83,19 @@ func DeserializePathElement(s string) (PathElement, error) {
 	case peKeySepBytes[0]:
 		iter := readPool.BorrowIterator(b)
 		defer readPool.ReturnIterator(iter)
-		v, err := value.ReadJSONIter(iter)
-		if err != nil {
-			return PathElement{}, err
-		}
-		if v.MapValue == nil {
-			return PathElement{}, fmt.Errorf("expected key value pairs but got %#v", v)
-		}
-		return PathElement{Key: v.MapValue}, nil
+
+		kv := []KeyValue{}
+		iter.ReadObjectCB(func(iter *jsoniter.Iterator, key string) bool {
+			v, err := value.ReadJSONIter(iter)
+			if err != nil {
+				iter.Error = err
+				return false
+			}
+			kv = append(kv, KeyValue{Key: key, Value: v})
+			return true
+		})
+		SortKeyValues(kv)
+		return PathElement{Key: &kv}, iter.Error
 	case peIndexSepBytes[0]:
 		i, err := strconv.Atoi(s[2:])
 		if err != nil {
@@ -129,13 +134,20 @@ func serializePathElementToWriter(w io.Writer, pe PathElement) error {
 		if _, err := stream.Write(peKeySepBytes); err != nil {
 			return err
 		}
-		v := value.Value{MapValue: pe.Key}
-		v.WriteJSONStream(stream)
+		stream.WriteObjectStart()
+		for i, kv := range *pe.Key {
+			if i > 0 {
+				stream.WriteMore()
+			}
+			stream.WriteObjectField(kv.Key)
+			value.WriteJSONStream(kv.Value, stream)
+		}
+		stream.WriteObjectEnd()
 	case pe.Value != nil:
 		if _, err := stream.Write(peValueSepBytes); err != nil {
 			return err
 		}
-		pe.Value.WriteJSONStream(stream)
+		value.WriteJSONStream(*pe.Value, stream)
 	case pe.Index != nil:
 		if _, err := stream.Write(peIndexSepBytes); err != nil {
 			return err