--- a/typed/helpers.go
+++ b/typed/helpers.go
@@ -117,11 +117,11 @@ func resolveSchema(s *schema.Schema, tr schema.TypeRef, v *value.Value, ah atomH
 func deduceAtom(a schema.Atom, v *value.Value) schema.Atom {
 	switch {
 	case v == nil:
-	case v.FloatValue != nil, v.IntValue != nil, v.StringValue != nil, v.BooleanValue != nil:
+	case value.IsFloat(*v), value.IsInt(*v), value.IsString(*v), value.IsBool(*v):
 		return schema.Atom{Scalar: a.Scalar}
-	case v.ListValue != nil:
+	case value.IsList(*v):
 		return schema.Atom{List: a.List}
-	case v.MapValue != nil:
+	case value.IsMap(*v):
 		return schema.Atom{Map: a.Map}
 	}
 	return a
@@ -149,90 +149,89 @@ func (ef errorFormatter) validateScalar(t *schema.Scalar, v *value.Value, prefix
 	if v == nil {
 		return nil
 	}
-	if v.Null {
+	if *v == nil {
 		return nil
 	}
 	switch *t {
 	case schema.Numeric:
-		if v.FloatValue == nil && v.IntValue == nil {
+		if !value.IsFloat(*v) && !value.IsInt(*v) {
 			// TODO: should the schema separate int and float?
-			return ef.errorf("%vexpected numeric (int or float), got %v", prefix, v)
+			return ef.errorf("%vexpected numeric (int or float), got %T", prefix, *v)
 		}
 	case schema.String:
-		if v.StringValue == nil {
-			return ef.errorf("%vexpected string, got %v", prefix, v)
+		if !value.IsString(*v) {
+			return ef.errorf("%vexpected string, got %#v", prefix, *v)
 		}
 	case schema.Boolean:
-		if v.BooleanValue == nil {
-			return ef.errorf("%vexpected boolean, got %v", prefix, v)
+		if !value.IsBool(*v) {
+			return ef.errorf("%vexpected boolean, got %v", prefix, *v)
 		}
 	}
 	return nil
 }
 
 // Returns the list, or an error. Reminder: nil is a valid list and might be returned.
-func listValue(val value.Value) (*value.List, error) {
-	switch {
-	case val.Null:
+func listValue(val value.Value) ([]interface{}, error) {
+	if val == nil {
 		// Null is a valid list.
 		return nil, nil
-	case val.ListValue != nil:
-		return val.ListValue, nil
-	default:
+	}
+	l, ok := val.([]interface{})
+	if !ok {
 		return nil, fmt.Errorf("expected list, got %v", val)
 	}
+	return l, nil
 }
 
 // Returns the map, or an error. Reminder: nil is a valid map and might be returned.
-func mapValue(val value.Value) (*value.Map, error) {
-	switch {
-	case val.Null:
+func mapValue(val value.Value) (value.Map, error) {
+	if val == nil {
+		// Null is a valid map.
 		return nil, nil
-	case val.MapValue != nil:
-		return val.MapValue, nil
-	default:
+	}
+	if !value.IsMap(val) {
 		return nil, fmt.Errorf("expected map, got %v", val)
 	}
+	return value.ValueMap(val), nil
 }
 
 func keyedAssociativeListItemToPathElement(list *schema.List, index int, child value.Value) (fieldpath.PathElement, error) {
 	pe := fieldpath.PathElement{}
-	if child.Null {
+	if child == nil {
 		// For now, the keys are required which means that null entries
 		// are illegal.
 		return pe, errors.New("associative list with keys may not have a null element")
 	}
-	if child.MapValue == nil {
+	if !value.IsMap(child) {
 		return pe, errors.New("associative list with keys may not have non-map elements")
 	}
-	keyMap := &value.Map{}
+	m := value.ValueMap(child)
+	keyValues := make([]fieldpath.KeyValue, 0, len(list.Keys))
 	for _, fieldName := range list.Keys {
-		var fieldValue value.Value
-		field, ok := child.MapValue.Get(fieldName)
-		if ok {
-			fieldValue = field.Value
+		if val, ok := m.Get(fieldName); ok {
+			keyValues = append(keyValues, fieldpath.KeyValue{fieldName, val})
 		} else {
-			// Treat keys as required.
 			return pe, fmt.Errorf("associative list with keys has an element that omits key field %q", fieldName)
 		}
-		keyMap.Set(fieldName, fieldValue)
 	}
-	pe.Key = keyMap
+	fieldpath.SortKeyValues(keyValues)
+	pe.Key = &keyValues
+
 	return pe, nil
 }
 
 func setItemToPathElement(list *schema.List, index int, child value.Value) (fieldpath.PathElement, error) {
 	pe := fieldpath.PathElement{}
 	switch {
-	case child.MapValue != nil:
+	case value.IsMap(child):
 		// TODO: atomic maps should be acceptable.
 		return pe, errors.New("associative list without keys has an element that's a map type")
-	case child.ListValue != nil:
+	case value.IsList(child):
 		// Should we support a set of lists? For the moment
 		// let's say we don't.
 		// TODO: atomic lists should be acceptable.
 		return pe, errors.New("not supported: associative list with lists as elements")
-	case child.Null:
+	case child == nil:
 		return pe, errors.New("associative list without keys has an element that's an explicit null")
 	default:
 		// We are a set type.