This is a WIP for a few reasons:
- Some things are much better, look at perf below,
- Some things are much worse, though it's still better overall,
- It moved some of the allocations from one place to another, I have ways to remove some altogether,
- I need to add more tests to value package
- This is based on the pathelement-set PR (#114)

Perf so far:
```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                              109043        91375         -16.20%
BenchmarkDeducedNested-12                              360530        336906        -6.55%
BenchmarkDeducedNestedAcrossVersion-12                 397451        370451        -6.79%
BenchmarkLeafConflictAcrossVersion-12                  94595         96292         +1.79%
BenchmarkMultipleApplierRecursiveRealConversion-12     1565003       1679940       +7.34%
BenchmarkOperations/Pod/Update-12                      100904        105945        +5.00%
BenchmarkOperations/Pod/Apply-12                       285154        270291        -5.21%
BenchmarkOperations/Node/Update-12                     158819        162062        +2.04%
BenchmarkOperations/Node/Apply-12                      494102        449763        -8.97%
BenchmarkOperations/Endpoints/Update-12                12728         15428         +21.21%
BenchmarkOperations/Endpoints/Apply-12                 1002018       1717987       +71.45%
BenchmarkFromUnstructured/Pod-12                       115741        34212         -70.44%
BenchmarkFromUnstructured/Node-12                      238225        69035         -71.02%
BenchmarkFromUnstructured/Endpoints-12                 6223548       1721100       -72.35%

benchmark                                              old allocs     new allocs     delta
BenchmarkDeducedSimple-12                              570            554            -2.81%
BenchmarkDeducedNested-12                              1910           1635           -14.40%
BenchmarkDeducedNestedAcrossVersion-12                 1978           1813           -8.34%
BenchmarkLeafConflictAcrossVersion-12                  505            540            +6.93%
BenchmarkMultipleApplierRecursiveRealConversion-12     5183           5045           -2.66%
BenchmarkOperations/Pod/Update-12                      255            399            +56.47%
BenchmarkOperations/Pod/Apply-12                       717            1032           +43.93%
BenchmarkOperations/Node/Update-12                     315            559            +77.46%
BenchmarkOperations/Node/Apply-12                      1076           1574           +46.28%
BenchmarkOperations/Endpoints/Update-12                45             69             +53.33%
BenchmarkOperations/Endpoints/Apply-12                 1076           7138           +563.38%
BenchmarkFromUnstructured/Pod-12                       370            161            -56.49%
BenchmarkFromUnstructured/Node-12                      854            312            -63.47%
BenchmarkFromUnstructured/Endpoints-12                 19076          7024           -63.18%
```