@apelisse 2019-10-09
I've added a couple of commits, the last one is definitely WIP since it's not tracking the pathelement (only used for errors, but we can refactor error handling to only build the path on errors).

These two commits significantly improve the situation though (all tests are still passing):
```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                              117303        91303         -22.16%
BenchmarkDeducedNested-12                              318361        337676        +6.07%
BenchmarkDeducedNestedAcrossVersion-12                 317505        352831        +11.13%
BenchmarkLeafConflictAcrossVersion-12                  115084        93126         -19.08%
BenchmarkMultipleApplierRecursiveRealConversion-12     1649080       1588821       -3.65%
BenchmarkOperations/Pod/Update-12                      351283        108252        -69.18%
BenchmarkOperations/Pod/Apply-12                       516962        253059        -51.05%
BenchmarkOperations/Node/Update-12                     569627        166258        -70.81%
BenchmarkOperations/Node/Apply-12                      755372        415201        -45.03%
BenchmarkOperations/Endpoints/Update-12                56002         15827         -71.74%
BenchmarkOperations/Endpoints/Apply-12                 2081751       30775         -98.52%
BenchmarkFromUnstructured/Pod-12                       202841        26335         -87.02%
BenchmarkFromUnstructured/Node-12                      327686        54673         -83.32%
BenchmarkFromUnstructured/Endpoints-12                 7909537       1097087       -86.13%

benchmark                                              old allocs     new allocs     delta
BenchmarkDeducedSimple-12                              572            543            -5.07%
BenchmarkDeducedNested-12                              1562           1588           +1.66%
BenchmarkDeducedNestedAcrossVersion-12                 1630           1766           +8.34%
BenchmarkLeafConflictAcrossVersion-12                  544            529            -2.76%
BenchmarkMultipleApplierRecursiveRealConversion-12     5268           4923           -6.55%
BenchmarkOperations/Pod/Update-12                      843            399            -52.67%
BenchmarkOperations/Pod/Apply-12                       1179           988            -16.20%
BenchmarkOperations/Node/Update-12                     1155           559            -51.60%
BenchmarkOperations/Node/Apply-12                      1876           1453           -22.55%
BenchmarkOperations/Endpoints/Update-12                100            69             -31.00%
BenchmarkOperations/Endpoints/Apply-12                 3143           127            -95.96%
BenchmarkFromUnstructured/Pod-12                       477            49             -89.73%
BenchmarkFromUnstructured/Node-12                      1107           109            -90.15%
BenchmarkFromUnstructured/Endpoints-12                 21114          7              -99.97%
```