@lavalamp 2019-10-23
so, overall, we currently are doing 2 conversions, typed struct -> unstructured -> value.Value.

This change removes the latter conversion, at some cost to this library's interface & functionality (?).

But to get rid of both conversions, the only choice is to make some sort of interface and some way of letting our existing typed structs implement it. In person Antoine said that iterating a map is expensive as the closure causes the variables to escape and be placed on the heap, making a lot of allocations. But without this I don't see how we can get rid of the 2nd conversion.

So, my current recommendation is:
* preserve the existing value type / package
* make an interface with the features of the current value.Value (which the existing value type will implement)
* make a wrapper that exposes map[string]interface{} in terms of this interface (intermediate checkpoint)
* make a wrapper that uses reflection to to expose arbitrary go objects in terms of this interface (2nd checkpoint - usually needed as a fallback for generated code)
* make a wrapper (or generate code, ugh) functions that expose our built-in resources in terms of this interface - this has to go in the main repo...

I admit to being slightly biased to like the existing value.Value because I think it is just a better version of an unstructured k8s object. An alternative to this PR is to remove the unstructured step and go directly from struct to value.Value in k8s code - basically swap out the internals of the unstructured object there.