@apelisse 2019-10-28
> - make an interface with the features of the current value.Value (which the existing value type will implement)
> - make a wrapper that exposes map[string]interface{} in terms of this interface (intermediate checkpoint)

I've started working on these points, but so far the results are unsatisfying. I'll quickly check for something obvious, but I won't have a lot of time for that. Unless we actually manage to make your suggested changes work, I would recommend we move forward with this PR.

```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkFromUnstructured/Pod-12           23063         43953         +90.58%
BenchmarkFromUnstructured/Node-12          47925         102539        +113.96%
BenchmarkFromUnstructured/Endpoints-12     865869        1764150       +103.74%

benchmark                                  old allocs     new allocs     delta
BenchmarkFromUnstructured/Pod-12           49             245            +400.00%
BenchmarkFromUnstructured/Node-12          109            625            +473.39%
BenchmarkFromUnstructured/Endpoints-12     7              9038           +129014.29%
```