@apelisse 2019-10-28
I just looked at pprof and confirmed with the escape analysis (`-gcflags '-m'`) that the extra wrapper is the thing that escapes to the heap.

I want to note that while I agree with your point in principle, and definitely believe that this is a possible further improvement, I think we should move forward with this existing PR code since it provides a significant improvement that we shouldn't give-up (and also given that there is no concern regarding correctness of this code).